Refactor code into sub-packages
