+++ src/main/java/duke/Main.java
@@ -3,6 +3,7 @@ package duke;
 import java.io.IOException;
 import java.io.InputStream;
 
+import duke.gui.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
