+++ src/main/java/duke/command/Commands.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.command;
 
 public class Commands {
     public enum CommandTypes {
