+++ src/main/java/duke/exception/DukeException.java
@@ -1,0 +1,15 @@
+package duke.exception;
+
+/**
+ * DukeException specifies the message to print when an exception is thrown in Duke.
+ *
+ * @author Gabriel Goh
+ */
+public class DukeException extends Exception {
+
+    public DukeException(String message) {
+        super(message);
+    }
+
+}
+
