+++ src/main/java/duke/exception/IllegalTaskTypeException.java
@@ -1,0 +1,12 @@
+package duke.exception;
+
+/**
+ * Throws exception when task type does not exist.
+ */
+public class IllegalTaskTypeException extends DukeException {
+
+    public IllegalTaskTypeException(String taskTypeStr) {
+        super(("Meow? There is no task type " + taskTypeStr + "!"));
+    }
+
+}
