+++ src/main/java/duke/exception/LoadFileMissing.java
@@ -1,0 +1,11 @@
+package duke.exception;
+
+/**
+ * Throws exception when load file does not exist.
+ */
+public class LoadFileMissing extends DukeException {
+
+    public LoadFileMissing() {
+        super("Meow? File to load does not exist.");
+    }
+}
