+++ src/main/java/duke/exception/MissingParams.java
@@ -1,0 +1,12 @@
+package duke.exception;
+
+/**
+ * Throws exception when expected parameters are missing.
+ */
+public class MissingParams extends DukeException {
+
+    public MissingParams(String params) {
+        super("Meow? Missing params... Please use /" + params + " params.");
+    }
+
+}
