+++ src/main/java/duke/exception/TaskIndexOutOfBounds.java
@@ -1,0 +1,12 @@
+package duke.exception;
+
+/**
+ * Throws exception when task is integer but is out of bounds of task list.
+ */
+public class TaskIndexOutOfBounds extends DukeException {
+
+    public TaskIndexOutOfBounds(int i, int length) {
+        super("Meow? I can't find task " + i + "... Please enter a task number between 1 and " + length);
+    }
+
+}
