+++ src/main/java/duke/parser/Parser.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.parser;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -17,6 +17,18 @@ import com.zoho.hawking.datetimeparser.configuration.HawkingConfiguration;
 import com.zoho.hawking.language.english.model.DatesFound;
 import com.zoho.hawking.language.english.model.ParserOutput;
 
+import duke.Duke;
+import duke.command.Commands;
+import duke.exception.DateTimeFormatException;
+import duke.exception.DukeException;
+import duke.exception.EmptyListException;
+import duke.exception.IllegalCommandException;
+import duke.exception.IllegalTaskTypeException;
+import duke.exception.MissingArguments;
+import duke.exception.MissingParams;
+import duke.exception.NothingAfterCommand;
+import duke.exception.TaskIndexNotInteger;
+import duke.task.Task;
 
 /**
  * Parser handles parsing of user input.
